type erasing for vulkan backend
