--- src/Layer/include/SkiaImpl/VSkiaEngine.h
@@ -1,9 +1,0 @@
-#pragma once
-
-#include "core/SkCanvas.h"
-#include "core/SkSurface.h"
-
-namespace VGG::layer::skia_impl
-{
-
-} // namespace VGG::layer::skia_impl
