--- src/Layer/include/SkiaImpl/VSkiaVK.h
@@ -1,3 +1,0 @@
-#pragma once
-
-#include "gpu/GrDirectContext.h"
